feat: Add server and client functionality (#master)
This commit introduces a server and a client that can communicate with each other. The server can accept connections and handle clients in separate threads. It can receive images from the client and add them to a list. The client can connect to the server, send images, and receive responses. An image processing logic has been added to handle the received images. A GUI for the server has been implemented using the Tkinter library. The commit also includes addition of metadata files with information about the author, version, license, and others.