feat: Upgrade nepyc package and add server functionality (#master)
Update the version of nepyc from 1.0.0-dev1 to 1.0.0-dev2. Include a
description for the project. Additionally, adjust the python
dependency to be compatible with versions >=3.10 and <3.14.

Introduce a script for nepyc-server. Add a dev dependency for
pyinstaller.

Create two new files, nepyc/server/main.py and
examples/server_example.py, to implement and exemplify the use of
ImageServer.

Finally, Introduce nepyc/main.py to the project.