feat: Upgrade nepyc package, add server functionality (#dev)
Update nepyc version from 1.0.0-dev1 to 1.0.0-dev2 and adjust python
compatibility to >=3.10 and <3.14. Introduce a script for nepyc-server
and add a dev dependency for pyinstaller.

New files nepyc/server/main.py and examples/server_example.py are
created to implement and exemplify ImageServer usage.

Additionally, nepyc/main.py is introduced to the project and .lock is
added to .gitignore.